tx. 1) Add improved CFF2 support: tx can now use the -U option to specify user design space coordinates snapshotting to a CFF or T1 font from a CFF2 source font. The CFF2 source data must in a complete OpenType font.  tx can now use the option -cff2 to write from a CFF2 font (in a complete OpenType font) to a file containing the output CFF2 table, with full charstring optimization and subroutinization. This is last option is still work in progress: it has not been extensively tested, and does yet support blended hints.
2) Updated subroutinizer to much faster with larger fonts. (by Ariza Michiharu)

3) Added new option (+V/-V) to remove overlaps.  (by Ariza Michiharu)

+++ FDK/Tools/Programs/public/lib/source/cffwrite/cffwrite_sindex.c
@@ -232,7 +232,7 @@ void cfwSindexWrite(cfwCtx g) {
 	index.count = (unsigned short)h->custom.cnt;
 	index.offSize = INDEX_OFF_SIZE(index.datasize);
 
-	cfwWrite2(g, index.count);
+	cfwWrite2(g, (unsigned short)index.count);
 
 	if (index.count == 0) {
 		return; /* Empty table just has zero count */
